Assignment 2/Source/requirements.txt
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.2
pycairo==1.20.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.18.1
scipy==1.6.1
six==1.15.0
tifffile==2021.3.17
